# openapi.json
{
  "components": {
    "schemas": {
      "AdjustIn": {
        "properties": {
          "direction": {
            "enum": [
              "grant",
              "redeem"
            ],
            "title": "Direction",
            "type": "string"
          },
          "note": {
            "maxLength": 200,
            "minLength": 3,
            "title": "Note",
            "type": "string"
          },
          "seconds": {
            "maximum": 86400.0,
            "minimum": 60.0,
            "title": "Seconds",
            "type": "integer"
          }
        },
        "required": [
          "direction",
          "seconds",
          "note"
        ],
        "title": "AdjustIn",
        "type": "object"
      },
...
      "ConsumptionIn": {
        "properties": {
          "app_label": {
            "maxLength": 80,
            "title": "App Label",
            "type": "string"
          },
          "client_event_id": {
            "maxLength": 100,
            "minLength": 8,
            "title": "Client Event Id",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "seconds": {
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Seconds",
            "type": "integer"
          }
        },
        "required": [
          "client_event_id",
          "seconds",
          "app_label"
        ],
        "title": "ConsumptionIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LedgerEntryOut": {
        "properties": {
          "delta_seconds": {
            "title": "Delta Seconds",
            "type": "integer"
          },
          "entry_type": {
            "title": "Entry Type",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          }
        },
        "required": [
          "delta_seconds",
          "entry_type",
          "note",
          "occurred_at"
        ],
        "title": "LedgerEntryOut",
        "type": "object"
      },
...
      "StandingOut": {
        "properties": {
          "balance_seconds": {
            "title": "Balance Seconds",
            "type": "integer"
          },
          "block_reason": {
            "enum": [
              "none",
              "no_balance",
              "daily_cap",
              "guardian_stale",
              "idle"
            ],
            "title": "Block Reason",
            "type": "string"
          },
          "daily_cap_seconds": {
            "title": "Daily Cap Seconds",
            "type": "integer"
          },
          "freeze_tokens": {
            "title": "Freeze Tokens",
            "type": "integer"
          },
          "idle_days": {
            "title": "Idle Days",
            "type": "integer"
          },
          "idle_days_allowed": {
            "title": "Idle Days Allowed",
            "type": "integer"
          },
          "playable_seconds": {
            "title": "Playable Seconds",
            "type": "integer"
          },
          "seconds_until_reset": {
            "title": "Seconds Until Reset",
            "type": "integer"
          },
          "spent_today_seconds": {
            "title": "Spent Today Seconds",
            "type": "integer"
          },
          "streak_current": {
            "title": "Streak Current",
            "type": "integer"
          }
        },
        "required": [
          "balance_seconds",
          "playable_seconds",
          "daily_cap_seconds",
          "spent_today_seconds",
          "idle_days",
          "idle_days_allowed",
          "block_reason",
          "seconds_until_reset",
          "streak_current",
          "freeze_tokens"
        ],
        "title": "StandingOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Kartu waktu bermain yang dicetak dari materi belajar",
    "title": "BrainXP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/subjects/{subject_id}/ledger": {
      "get": {
        "operationId": "history_subjects__subject_id__ledger_get",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LedgerEntryOut"
                  },
                  "title": "Response History Subjects  Subject Id  Ledger Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "ledger"
        ]
      }
    },
    "/subjects/{subject_id}/ledger/adjust": {
      "post": {
        "operationId": "adjust_subjects__subject_id__ledger_adjust_post",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adjust",
        "tags": [
          "ledger"
        ]
      }
    },
    "/subjects/{subject_id}/ledger/sync": {
      "post": {
        "operationId": "sync_consumption_subjects__subject_id__ledger_sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ConsumptionIn"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Consumption",
        "tags": [
          "ledger"
        ]
      }
    },
...
    "/subjects/{subject_id}/standing": {
      "get": {
        "operationId": "standing_subjects__subject_id__standing_get",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Standing",
        "tags": [
          "ledger"
        ]
      }
    }
  }
}